added floating point

+++ calc.py
@@ -2,8 +2,8 @@ def compute (string):
    """Perform simple math based on string input. 
       Example '5+7' -> 12"""
     values =string.split(' ')
-	num0= int(values[0])
-	num1=int(values[2])
+	num0 = float(values[0])
+	num1 = float(values[2])
 	operator=values[1]
 	if operator =='+':
 	     return num0 +num1
